Migrate `update_path` to `UpdateHalfEdgeGeometry`
